backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
redis
httpx
pytest
pytest-asyncio
python-socketio[asyncio]
supabase
asyncpg
prometheus-fastapi-instrumentator
fastapi-limiter
bleach
celery
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
sentry-sdk[fastapi]
